♻️ refactor: split ChatPage and Sidebar into separate components with dedicated styles
