+++ src/pages/ChatPage.jsx
@@ -1,8 +1,7 @@
 import { useState, useEffect } from "react";
 import { getProfile } from "../api";
-import { clearToken } from "../auth";
-import { LogOut } from "lucide-react";
 import ChatBox from "../components/ChatBox";
+import Sidebar from "../components/Sidebar";
 import "../styles/ChatPage.css";
 
 export default function ChatPage() {
@@ -18,10 +17,7 @@ export default function ChatPage() {
         }
       } catch (err) {
         if (err.name === "ApiError" && err.status === 401) {
-          alert("⚠️ Session expired, redirecting to login...");
           window.location.href = "/login";
-        } else {
-          alert(`❌ ${err.message}`);
         }
       }
     };
@@ -36,66 +32,9 @@ export default function ChatPage() {
     );
   }
 
-  const getAvatarUrl = (user) => {
-    if (user.picture) return user.picture;
-    return `https://ui-avatars.com/api/?name=${encodeURIComponent(
-      user.name || "User"
-    )}&background=0D8ABC&color=fff&rounded=true`;
-  };
-
-  const currentYear = new Date().getFullYear();
-
   return (
     <div className="chat-layout">
-      {/* Sidebar */}
-      <aside className={`sidebar ${sidebarOpen ? "open" : "closed"}`}>
-        <div className="brand">🌿 Ranting AI</div>
-
-        <div className="sidebar-content">
-          <p className="comingsoon">+ New Chat (soon)</p>
-        </div>
-
-        <div className="profile">
-          <img
-            className="avatar"
-            src={getAvatarUrl(user)}
-            alt={user.name}
-            onError={(e) => {
-              e.currentTarget.src = `https://ui-avatars.com/api/?name=${encodeURIComponent(
-                user.name || "User"
-              )}&background=0D8ABC&color=fff&rounded=true`;
-            }}
-          />
-          <div className="user-info">
-            <span className="user-name">{user.name}</span>
-            <span className="user-email">{user.email}</span>
-          </div>
-          <button
-            className="logout-btn"
-            onClick={() => {
-              clearToken();
-              window.location.href = "/";
-            }}
-            title="Logout"
-          >
-            <LogOut size={15} />
-          </button>
-        </div>
-
-        <footer className="sidebar-footer">
-          Powered by{" "}
-          <a
-            href="https://akarkode.com"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Akarkode
-          </a>{" "}
-          © {currentYear}
-        </footer>
-      </aside>
-
-      {/* Main */}
+      <Sidebar open={sidebarOpen} user={user} />
       <div className="main">
         <header className="chat-header">
           <button
@@ -107,7 +46,6 @@ export default function ChatPage() {
           </button>
           <h1>Ranting AI Assistant</h1>
         </header>
-
         <main className="chat-main">
           <ChatBox />
         </main>
