+++ src/styles/ChatPage.css
@@ -3,89 +3,6 @@
   height: 100vh;
   overflow: hidden;
 }
-
-/* Sidebar */
-.sidebar {
-  width: 260px;
-  background: #111827;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  transition: width 0.3s ease-in-out, opacity 0.3s ease-in-out;
-  overflow: hidden;
-}
-.sidebar.closed {
-  width: 0;
-  opacity: 0;
-}
-.brand {
-  font-weight: 700;
-  font-size: 20px;
-  padding: 16px;
-  background: linear-gradient(90deg, #60a5fa, #c084fc);
-  background-clip: text;         /* standar */
-  -webkit-background-clip: text; /* Safari */
-  color: transparent;
-  -webkit-text-fill-color: transparent;
-}
-.sidebar-content {
-  flex: 1;
-  padding: 16px;
-  font-size: 14px;
-  color: #94a3b8;
-}
-.profile {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding: 16px;
-  border-top: 1px solid #1e293b;
-}
-.avatar {
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  object-fit: cover;
-  border: 2px solid #334155;
-}
-.user-info {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-}
-.user-name {
-  font-size: 14px;
-  font-weight: 600;
-}
-.user-email {
-  font-size: 12px;
-  color: #94a3b8;
-}
-.logout-btn {
-  border: none;
-  background: transparent;
-  color: #ef4444;
-  cursor: pointer;
-  padding: 6px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-/* Footer */
-.sidebar-footer {
-  margin-top: auto;
-  font-size: 12px;
-  color: #ffffff;
-  text-align: center;
-  padding: 10px;
-}
-.sidebar-footer a {
-  color: #ffffff;
-  text-decoration: underline;
-}
-
-/* Main */
 .main {
   flex: 1;
   display: flex;
@@ -120,8 +37,6 @@
   color: #fff;
   cursor: pointer;
 }
-
-/* Loading state */
 .loading {
   flex: 1;
   display: flex;
@@ -129,20 +44,4 @@
   align-items: center;
   color: #94a3b8;
   font-size: 16px;
-}
-
-/* Responsive */
-@media (max-width: 768px) {
-  .sidebar {
-    position: absolute;
-    height: 100%;
-    z-index: 20;
-  }
-  .sidebar.closed {
-    transform: translateX(-100%);
-    width: 260px; /* keep width but hidden */
-  }
-  .main {
-    flex: 1;
-  }
 }
